Added 3 methods to ProductModel which allow to: 1. get product by its id; 2. get all products; 3. get products by their type;
Added management page. In which created 'add new Product' and 'add new ProductType' buttons and 2 gridviews with the ability to edit and delete data about products(first gridview) and product types(second).

Added a method in management back code which allows when 'edit' button hit to redirect the page to manageProducts page. Makes data editing more comfy.

When ManageProducts page loads, I added checking if there's product id is present:
- If present, page will be filled with product info according to its id. It allows to edit info of existing product;
If not, user would be able to add a new product to submit;

In MasterPage I added NavigateUrl to Management link so when it(Management link) clicked opens Management page.